pyproject.toml
[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ask_ai.settings"
env = ["ENVIRONMENT=TEST",]

[tool.poetry]
name = "ask-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.2"
automatilib = "^1.2.1"
django = "^4.2.8"
whitenoise = "^6.6.0"
watchdog = {extras = ["watchmedo"], version = "^3.0.0"}
jinja2 = "^3.1.2"
django-environ = "^0.11.2"
django-use-email-as-username = "^1.4.0"
markdown-it-py = "^3.0.0"
waitress = "^2.1.2"
psycopg2-binary = "^2.9.9"
pyyaml = "^6.0.1"
openai = "0.28.1"
pytz = "^2023.3.post1"
django-permissions-policy = "^4.18.0"
django-csp = "^3.7"
django-single-session = "^0.2.0"
sentry-sdk = "^1.34.0"
django-storages = "^1.14.2"
django-health-check = "<3.18.0" # this pin can be removed once https://github.com/revsys/django-health-check/issues/412 is resolved
boto3 = "^1.33.10"
urllib3 = "<1.27"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.31.0"
tiktoken = "^0.5.1"
presidio-analyzer = "^2.2.351"
en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz"}
defusedcsv = "^2.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
freezegun = "^1.2.2"
pep8-naming = "^0.13.3"
bandit = "^1.7.5"
pytest = "^7.4.3"
pytest-django = "^4.6.0"
pytest-env = "^1.1.1"
pytest-mock = "^3.12.0"
mypy = "^1.6.1"
django-types = "^0.19.1"
types-python-jose = "^3.3.4.8"
pytest-cov = "^4.1.0"
detect-secrets = "^1.4.0"
pre-commit = "^3.6.0"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
